refactor(android): bring getSize to image adapter for custom-impl
